% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{measure}
\alias{measure}
\alias{unipromeasure}
\alias{maxpromeasure}
\alias{maximinmeasure}
\title{measure}
\usage{
measure(X, s = nrow(X), p = 15L, criterion = c("UniPro", "MaxPro", "maximinLHD"))

unipromeasure(X)

maxpromeasure(X)

maximinmeasure(X)
}
\arguments{
\item{s}{number of levels. Default = n}

\item{p}{the exponent for method maximinLHD. default is 15}

\item{method}{The criterion to be optimized. Only UniPro, MaxPro, maximinLHD are implemented}

\item{X}{Design}
}
\value{
measure value.
}
\description{
Computes measure for a given design criterion.

A function that computes the UniPro criterion value for a balanced design.
Equation 5.1 and 5.2 of the Uniform Projection Designs paper by Sun, Wang and Xu.

A function that computes the MaxPro criterion value for a given design.
Equation 5 of the  Maximum projection designs for computer experiments
paper BY V. ROSHAN JOSEPH,EVRENGUL

A function that computes the maximin criterion value for a given design.
(Morris and Mitchell 1995)
}
\examples{
measure(replicate(3, sample(10)))

unipromeasure(replicate(3, sample(10)))

maxpromeasure(replicate(3, sample(10)))

maximinmeasure(replicate(3, sample(10)))

}
