% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{psi}
\alias{psi}
\title{psi}
\usage{
psi(X, s = nrow(X))
}
\arguments{
\item{X}{Design}

\item{s}{number of levels for the design X}
}
\value{
the phi value.
}
\description{
A function that computes the MaxPro criterion value for a given design.
Equation 5 of the  Maximum projection designs for computer experiments
paper BY V. ROSHAN JOSEPH,EVRENGUL
}
\examples{
phi(replicate(3, sample(10)))

}
